moved the code, not well tested yet
