+++ src/open62541_compat.cpp
@@ -149,140 +149,8 @@ UaString UaDateTime::toString() const
 	return UaString(result.str().c_str());
 }
 
-UaDataValue::UaDataValue( const UaVariant& variant, OpcUa_StatusCode statusCode, const UaDateTime& serverTime, const UaDateTime& sourceTime ):
-    m_lock()
-{
-	m_lock.clear();
-    m_impl = UA_DataValue_new ();
-    if (!m_impl)
-        throw std::runtime_error( "UA_DataValue_new returned 0" );
-    // LOG(Log::INF) << "allocated new UA_DataValue @ " <<  m_impl;  
-
-    // TODO: duplicate the variant
-    UA_Variant_copy( variant.impl(), &m_impl->value );
-    LOG(Log::TRC) << "After UA_Variant_copy: src="<<variant.impl()<<" src.data="<<variant.impl()->data<<" dst="<<&m_impl->value<<" dst.data="<<m_impl->value.data;
-
-    m_impl->hasValue = 1;
-
-}
-
-UaDataValue::UaDataValue( const UaDataValue& other ):
-    m_lock()
-{
-	m_lock.clear();
-    m_impl = UA_DataValue_new ();
-    // LOG(Log::INF) << "allocated new UA_DataValue @ " <<  m_impl;  
-    UA_DataValue_copy( other.m_impl, m_impl );
-}
-
-void UaDataValue:: operator=(const UaDataValue& other )
-{
-    while (m_lock.test_and_set(std::memory_order_acquire));  // acquire lock
-    if (m_impl)
-    {
-        UA_DataValue_deleteMembers (m_impl);
-        UA_DataValue_delete( m_impl );
-        m_impl = 0;
-    }
-    m_impl = UA_DataValue_new ();
-    // LOG(Log::INF) << "allocated new UA_DataValue @ " <<  m_impl;  
-    UA_DataValue_copy( other.m_impl, m_impl );
-    m_lock.clear(std::memory_order_release);
-    
-}
-
-UaDataValue UaDataValue::clone()
-{
-    while (m_lock.test_and_set(std::memory_order_acquire));  // acquire lock
-    UaDataValue aCopy ( *this );
-    m_lock.clear(std::memory_order_release);
-    return aCopy;
-    
-}
-
-UaDataValue:: ~UaDataValue ()
-{
-    if (m_impl)
-    {
-	UA_DataValue_deleteMembers( m_impl );
-	UA_DataValue_delete( m_impl );
-	m_impl = 0;
-    }
-    
-}
-
-UaNodeId::UaNodeId ( const UaString& stringAddress, int ns)
-{
-    // TODO: not implemented yet in open62541
-    //    UA_NodeId_fromInteger( 2, 2);
-    m_impl.namespaceIndex = ns;
-    m_impl.identifierType = UA_NODEIDTYPE_STRING;
-    UA_StatusCode status = UA_String_copy( stringAddress.impl(), &m_impl.identifier.string );
-    if (status != UA_STATUSCODE_GOOD)
-        throw alloc_error();
-}
-
-
-UaNodeId::UaNodeId( int numericAddress, int ns ):
-    m_impl( UA_NODEID_NUMERIC( ns, numericAddress ))
-{
-    
-}
-
-UaNodeId::UaNodeId ( const UaNodeId& other)
-{
-    UA_NodeId_init( &m_impl );
-    UA_StatusCode status = UA_NodeId_copy( other.pimpl(), &this->m_impl );
-    if (status != UA_STATUSCODE_GOOD)
-        throw alloc_error();
-
-}
-
-UaNodeId::~UaNodeId ()
-{
-    UA_NodeId_deleteMembers( &m_impl );
-}
-
-IdentifierType UaNodeId::identifierType() const
-{
-    switch (m_impl.identifierType)
-    {
-    case UA_NODEIDTYPE_NUMERIC: return IdentifierType::OpcUa_IdentifierType_Numeric;
-    case UA_NODEIDTYPE_STRING: return IdentifierType::OpcUa_IdentifierType_String; 
-    default: throw std::runtime_error("not-implemented");
-    }
-}
-
-UaString UaNodeId::identifierString() const 
-{
-    if (m_impl.identifierType != UA_NODEIDTYPE_STRING)
-        throw std::runtime_error("asking for identifierString from a non-string identifier!");
-    return UaString( &m_impl.identifier.string );
-
-}	     
-
-bool UaNodeId::operator==(const UaNodeId& other) const
-{
-    return UA_NodeId_equal( &m_impl, &other.m_impl );
-}
-
 UaNode::UaNode ()
 {
-}
-
-UaString UaNodeId::toString() const
-{
-    if (identifierType() == IdentifierType::OpcUa_IdentifierType_String)
-    {
-        std::string s = "(ns="+boost::lexical_cast<std::string>(namespaceIndex())+","+UaString(identifierString()).toUtf8()+")";
-        return UaString(s.c_str());
-    }
-    else if (identifierType() == IdentifierType::OpcUa_IdentifierType_Numeric)
-    {
-        std::string s = "(ns="+boost::lexical_cast<std::string>(namespaceIndex())+","+boost::lexical_cast<std::string>(identifierNumeric())+")";
-        return UaString(s.c_str());
-    }
-    return "non-string-id";
 }
 
 namespace OpcUa
